# pkg/latex/docs/swagger.yaml
basePath: /api/doc
info:
  contact:
    email: support@bullteam.cn
    name: API Support
  description: TEFS latex
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TEFS latex API
  version: V0.1
paths:
...
  /latex/{id}/mv:
    get:
      consumes:
      - application/json
      description: 移动latex文件或目录
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: 源路径
        in: query
        name: source_path
        required: true
        type: string
      - description: 目标路径
        in: query
        name: target_path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 移动latex文件或目录
      tags:
      - latex
...
  /latex/{id}/tree:
    get:
      consumes:
      - application/json
      description: 获取latex目录树
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取latex目录树
      tags:
      - latex
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"